Update the hex sequence for option-left and option-right
The hex sequence has a double-escape which confuses nushell (but zsh is
fine). If we change it to ^[b and ^[f then it would correspond to
"back" and "forward" respectively, which performs the right action but
could be ambiguous as to which key the user pressed.

Changed due to me trying out nushell.